refactor: extract anomaly application logic into dedicated classes
